FRUSTRATION
So Frustrated. Something is wrong with linking to Magick++. It's really stupid.
I'm doing something wrong but I can't figure it out. I'm so frustrated.
It complains that it can't find "Magick::Images" but It can find "Magick::Blob".
Wut. That makes no sense. I'm linking it to the same fucking libraries.
What's weirder is that it compiles in my other project. I can't figure out the
difference between the two. Fuck my life.